Remove old repositories
